# cmd/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateExchangeRateRequest:
    properties:
      dateEffective:
        type: string
      fromCurrencyCode:
        type: string
      rate:
        description: Consider adding validation for > 0
        type: number
      toCurrencyCode:
        type: string
    required:
    - dateEffective
    - fromCurrencyCode
    - rate
    - toCurrencyCode
    type: object
...
  dto.ExchangeRateResponse:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      dateEffective:
        type: string
      exchangeRateID:
        type: string
      fromCurrencyCode:
        type: string
      lastUpdatedAt:
        type: string
      lastUpdatedBy:
        type: string
      rate:
        type: number
      toCurrencyCode:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for MMA backend.
  title: MMA Backend API
  version: "1.0"
paths:
...
  /exchange-rates:
    post:
      consumes:
      - application/json
      description: Adds a new exchange rate between two currencies for a specific
        date
      parameters:
      - description: Exchange Rate details
        in: body
        name: rate
        required: true
        schema:
          $ref: '#/definitions/dto.CreateExchangeRateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ExchangeRateResponse'
        "400":
          description: Invalid input format or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create exchange rate
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new exchange rate
      tags:
      - exchange rates
  /exchange-rates/{from}/{to}:
    get:
      description: Retrieves the latest exchange rate for a given currency pair
      parameters:
      - description: From Currency Code (3 letters)
        in: path
        maxLength: 3
        minLength: 3
        name: from
        required: true
        type: string
      - description: To Currency Code (3 letters)
        in: path
        maxLength: 3
        minLength: 3
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ExchangeRateResponse'
        "400":
          description: Invalid currency code format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Exchange rate not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve exchange rate
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an exchange rate
      tags:
      - exchange rates
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"